chore: refactor structure of mgmt related form file
