+++ src/components/forms/index.ts
@@ -1,7 +1,4 @@
-import OnboardingForm from "./OnboardingForm";
-
 export * from "./connector";
 export * from "./model";
 export * from "./pipeline";
-
-export { OnboardingForm };
+export * from "./mgmt";
