+++ src/components/forms/mgmt/index.ts
@@ -1,0 +1,8 @@
+import OnboardingForm from "./OnboardingForm";
+import type {
+  OnBoardingFormProps,
+  OnboardingFormValue,
+} from "./OnboardingForm";
+
+export { OnboardingForm };
+export type { OnBoardingFormProps, OnboardingFormValue };
